+++ BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeClient.java
@@ -27,7 +27,7 @@ public class BraintreeClient {
     private final String sessionId;
     private final String integrationType;
 
-    private static BraintreeClientParams createDefaultParams(Context context, String authString) throws InvalidArgumentException {
+    private static BraintreeClientParams createDefaultParams(Context context, String authString) {
         Authorization authorization = Authorization.fromString(authString);
         BraintreeHttpClient httpClient = new BraintreeHttpClient(authorization);
         return new BraintreeClientParams()
@@ -47,10 +47,9 @@ public class BraintreeClient {
      * Create a new instance of {@link BraintreeClient} using a tokenization key or client token.
      *
      * @param context       Android Context
-     * @param authorization The tokenization key or client token to use.
-     * @throws InvalidArgumentException If the tokenization key or client token is not valid, or cannot be parsed.
+     * @param authorization The tokenization key or client token to use. If an invalid authorization is provided, a {@link BraintreeException} will be returned in the {@link HttpResponseCallback}.
      */
-    public BraintreeClient(Context context, String authorization) throws InvalidArgumentException {
+    public BraintreeClient(Context context, String authorization) {
         this(createDefaultParams(context, authorization));
     }