+++ BraintreeCore/src/main/java/com/braintreepayments/api/BraintreeHttpClient.java
@@ -68,6 +68,12 @@ class BraintreeHttpClient {
      * @param retryStrategy retry strategy
      */
     void get(String path, Configuration configuration, @RetryStrategy int retryStrategy, HttpResponseCallback callback) {
+        if (authorization instanceof InvalidAuthorization) {
+            String message = ((InvalidAuthorization) authorization).getErrorMessage();
+            callback.failure(new BraintreeException(message));
+            return;
+        }
+
         boolean isRelativeURL = !path.startsWith("http");
         if (configuration == null && isRelativeURL) {
             String message = "Braintree HTTP GET request without configuration cannot have a relative path.";
@@ -112,6 +118,12 @@ class BraintreeHttpClient {
      * @param configuration configuration for the Braintree Android SDK.
      */
     void post(String path, String data, Configuration configuration, HttpResponseCallback callback) {
+        if (authorization instanceof InvalidAuthorization) {
+            String message = ((InvalidAuthorization) authorization).getErrorMessage();
+            callback.failure(new BraintreeException(message));
+            return;
+        }
+
         boolean isRelativeURL = !path.startsWith("http");
         if (configuration == null && isRelativeURL) {
             String message = "Braintree HTTP GET request without configuration cannot have a relative path.";
@@ -160,6 +172,11 @@ class BraintreeHttpClient {
      * @return the HTTP response body
      */
     String post(String path, String data, Configuration configuration) throws Exception {
+        if (authorization instanceof InvalidAuthorization) {
+            String message = ((InvalidAuthorization) authorization).getErrorMessage();
+            throw new BraintreeException(message);
+        }
+
         boolean isRelativeURL = !path.startsWith("http");
         if (configuration == null && isRelativeURL) {
             String message = "Braintree HTTP GET request without configuration cannot have a relative path.";