+++ BraintreeModels/src/main/java/com/braintreepayments/api/InvalidAuthorization.java
@@ -1,0 +1,25 @@
+package com.braintreepayments.api;
+
+class InvalidAuthorization extends Authorization {
+
+    private final String errorMessage;
+
+    InvalidAuthorization(String rawValue, String errorMessage) {
+        super(rawValue);
+        this.errorMessage = errorMessage;
+    }
+
+    @Override
+    String getConfigUrl() {
+        return null;
+    }
+
+    @Override
+    String getBearer() {
+        return null;
+    }
+
+    String getErrorMessage() {
+        return errorMessage;
+    }
+}