+++ BraintreeModels/src/test/java/com/braintreepayments/api/PayPalUATUnitTest.java
@@ -1,22 +1,19 @@
 package com.braintreepayments.api;
 
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import java.util.Base64;
 
 import static junit.framework.Assert.assertEquals;
+import static junit.framework.Assert.assertTrue;
 
 @RunWith(RobolectricTestRunner.class)
 public class PayPalUATUnitTest {
-    @Rule
-    public ExpectedException exceptionRule = ExpectedException.none();
 
     @Test
-    public void fromString_setsAllProperties() throws InvalidArgumentException {
+    public void fromString_setsAllProperties() {
         PayPalUAT payPalUAT = (PayPalUAT) Authorization.fromString(Fixtures.BASE64_PAYPAL_UAT);
 
         assertEquals("https://api.sandbox.braintreegateway.com:443/merchants/cfxs3ghzwfk2rhqm/client_api/v1/configuration", payPalUAT.getConfigUrl());
@@ -28,14 +25,14 @@ public class PayPalUATUnitTest {
     // "iss" field properly indicates env
 
     @Test
-    public void fromString_withStagingIssuer_setsProperEnv() throws InvalidArgumentException {
+    public void fromString_withStagingIssuer_setsProperEnv() {
         PayPalUAT payPalUAT = (PayPalUAT) Authorization.fromString(encodeUAT("{\"iss\":\"https://api.msmaster.qa.paypal.com\", \"external_id\":[\"Braintree:id\"]}"));
 
         assertEquals(PayPalUAT.Environment.STAGING, payPalUAT.getEnvironment());
     }
 
     @Test
-    public void fromString_withSandboxIssuer_setsProperEnv() throws InvalidArgumentException {
+    public void fromString_withSandboxIssuer_setsProperEnv() {
         PayPalUAT payPalUAT = (PayPalUAT) Authorization.fromString(encodeUAT("{\"iss\":\"https://api.paypal.com\", \"external_id\":[\"Braintree:id\"]}"));
 
         assertEquals(PayPalUAT.Environment.PRODUCTION, payPalUAT.getEnvironment());
@@ -44,42 +41,46 @@ public class PayPalUATUnitTest {
     // error scenarios
 
     @Test
-    public void fromString_withMalformedUAT_throwsException() throws InvalidArgumentException {
-        exceptionRule.expect(InvalidArgumentException.class);
-        exceptionRule.expectMessage("Authorization provided is invalid");
+    public void fromString_withMalformedUAT_returnsAnInvalidToken() {
+        Authorization result = PayPalUAT.fromString("invalid.uat-without-signature");
 
-        PayPalUAT.fromString("invalid.uat-without-signature");
+        assertTrue(result instanceof InvalidAuthorization);
+        String expectedErrorMessage = "Authorization provided is invalid: invalid.uat-without-signature";
+        assertEquals(expectedErrorMessage, ((InvalidAuthorization) result).getErrorMessage());
     }
 
     @Test
-    public void fromString_whenJSONSerializationFails_throwsException() throws Exception {
-        exceptionRule.expect(InvalidArgumentException.class);
+    public void fromString_whenJSONSerializationFails_returnsAnInvalidToken() {
+        Authorization result = PayPalUAT.fromString(encodeUAT("{\"some_invalid_json\": "));
 
-        PayPalUAT.fromString(encodeUAT("{\"some_invalid_json\": "));
+        assertTrue(result instanceof InvalidAuthorization);
     }
 
     @Test
-    public void fromString_whenNoBraintreeMerchantID_throwsException() throws Exception {
-        exceptionRule.expect(InvalidArgumentException.class);
-        exceptionRule.expectMessage("PayPal UAT invalid: Missing Braintree merchant account ID.");
+    public void fromString_whenNoBraintreeMerchantID_returnsAnInvalidToken() {
+        Authorization result = PayPalUAT.fromString(encodeUAT("{\"iss\":\"paypal-url.com\", \"external_id\":[\"Faketree:my-merchant-id\"]}"));
 
-        PayPalUAT.fromString(encodeUAT("{\"iss\":\"paypal-url.com\", \"external_id\":[\"Faketree:my-merchant-id\"]}"));
+        assertTrue(result instanceof InvalidAuthorization);
+        String expectedErrorMessage = "PayPal UAT invalid: Missing Braintree merchant account ID.";
+        assertEquals(expectedErrorMessage, ((InvalidAuthorization) result).getErrorMessage());
     }
 
     @Test
-    public void fromString_whenNoIssuerPresent_throwsException() throws Exception {
-        exceptionRule.expect(InvalidArgumentException.class);
-        exceptionRule.expectMessage("PayPal UAT invalid: Does not contain issuer, or \"iss\" key.");
+    public void fromString_whenNoIssuerPresent_returnsAnInvalidToken() {
+        Authorization result = PayPalUAT.fromString(encodeUAT("{\"external_id\":[\"Braintree:my-merchant-id\"]}"));
 
-        PayPalUAT.fromString(encodeUAT("{\"external_id\":[\"Braintree:my-merchant-id\"]}"));
+        assertTrue(result instanceof InvalidAuthorization);
+        String expectedErrorMessage = "PayPal UAT invalid: Does not contain issuer, or \"iss\" key.";
+        assertEquals(expectedErrorMessage, ((InvalidAuthorization) result).getErrorMessage());
     }
 
     @Test
-    public void fromString_whenPayPalURLUnknown_throwsException() throws Exception {
-        exceptionRule.expect(InvalidArgumentException.class);
-        exceptionRule.expectMessage("PayPal issuer URL missing or unknown: fake-url.com");
+    public void fromString_whenPayPalURLUnknown_returnsAnInvalidToken() {
+        Authorization result = PayPalUAT.fromString(encodeUAT("{\"iss\":\"fake-url.com\", \"external_id\":[\"Braintree:my-merchant-id\"]}"));
 
-        PayPalUAT.fromString(encodeUAT("{\"iss\":\"fake-url.com\", \"external_id\":[\"Braintree:my-merchant-id\"]}"));
+        assertTrue(result instanceof InvalidAuthorization);
+        String expectedErrorMessage = "PayPal UAT invalid: PayPal issuer URL missing or unknown: fake-url.com";
+        assertEquals(expectedErrorMessage, ((InvalidAuthorization) result).getErrorMessage());
     }
 
     // Test Helpers